backend/requirements.txt
Flask
gunicorn
flask-cors
openai
python-dotenv
requests
pydub
ffmpeg-python
moviepy
python-docx
pydantic
PyPDF2
kokoro>=0.3.4
soundfile
